@@ -1,12 +1,12 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 25
-    buildToolsVersion "25.0.2"
+    compileSdkVersion 33
+    buildToolsVersion "31.0.0"
     defaultConfig {
         applicationId "io.smileyjoe.applist"
         minSdkVersion 21
-        targetSdkVersion 25
+        targetSdkVersion 33
         versionCode 3
         versionName "1.2"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
@@ -18,22 +18,23 @@ android {
         }
 
         debug {
-            buildConfigField "String", "SERVER_CLIENT_ID", appListServerClientId
+            applicationIdSuffix ".debug"
+            buildConfigField "String", "SERVER_CLIENT_ID", findProperty("app_list.firebase_server_client_id")
         }
     }
 }
 
 dependencies {
-    compile fileTree(include: ['*.jar'], dir: 'libs')
-    androidTestCompile('com.android.support.test.espresso:espresso-core:2.2.2', {
+    implementation fileTree(include: ['*.jar'], dir: 'libs')
+    androidTestImplementation('com.android.support.test.espresso:espresso-core:2.2.2', {
         exclude group: 'com.android.support', module: 'support-annotations'
     })
-    compile 'com.android.support:appcompat-v7:25.3.1'
-    compile 'com.android.support:design:25.3.1'
-    compile 'za.co.smileyjoedev.lib:debug:1.0.0'
-    compile 'com.google.firebase:firebase-database:10.2.1'
-    testCompile 'junit:junit:4.12'
-    compile project(':firebaseauth')
+    implementation 'androidx.appcompat:appcompat:1.5.1'
+    implementation 'com.google.android.material:material:1.7.0'
+    implementation 'com.google.firebase:firebase-database:10.2.1'
+    implementation 'com.google.firebase:firebase-auth:21.1.0'
+    testImplementation 'junit:junit:4.12'
+    implementation project(':firebaseauth')
 }
 
 apply plugin: 'com.google.gms.google-services'
