@@ -1,8 +1,9 @@
 package io.smileyjoe.applist.activity;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
 
 import io.smileyjoe.applist.object.User;
 
