@@ -1,9 +1,10 @@
 package io.smileyjoe.applist.adapter;
 
 import android.content.Context;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentPagerAdapter;
+
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentPagerAdapter;
 
 import io.smileyjoe.applist.fragment.AppListFragment;
 import io.smileyjoe.applist.fragment.AppListFragment.Type;
