@@ -5,6 +5,7 @@ import android.app.Activity;
 import com.google.firebase.database.DatabaseReference;
 import com.google.firebase.database.FirebaseDatabase;
 
+import io.smileyjoe.applist.BuildConfig;
 import io.smileyjoe.applist.R;
 import io.smileyjoe.applist.object.User;
 
@@ -14,7 +15,7 @@ import io.smileyjoe.applist.object.User;
 
 public class Db {
 
-    private static final String DB_KEY_APP_DETAIL = "app";
+    private static final String DB_KEY_APP_DETAIL = BuildConfig.DEBUG ? "app-debug" : "app";
 
     private Db() {
     }
