@@ -1,10 +1,11 @@
 package io.smileyjoe.applist.viewholder;
 
-import android.support.v7.widget.RecyclerView;
 import android.view.View;
 import android.widget.Button;
 import android.widget.ImageView;
 import android.widget.TextView;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 import io.smileyjoe.applist.R;
 import io.smileyjoe.applist.fragment.AppListFragment;
